backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]
pydantic
python-dotenv
pymongo
passlib[bcrypt]
bcrypt==3.2.0
python-jose
pydantic[email]
motor
SQLAlchemy
python-multipart
beanie
celery
redis
openai
tavily-python
PyJWT